+++ AI/VCAI/Goals.cpp
@@ -10,7 +10,7 @@
 #include "StdInc.h"
 #include "Goals.h"
 #include "VCAI.h"
-#include "Fuzzy.h"
+#include "FuzzyHelper.h"
 #include "ResourceManager.h"
 #include "BuildingManager.h"
 #include "../../lib/mapping/CMap.h" //for victory conditions
